feat: turn few admin actions into tasks
